feat(FormModel): remove currentOwner parameter from Form#transferOwner
Not needed, since the form the instance method is acting on should already have the required information, and should only have a single source of truth.

The issue where the current logged in user calling this method to transfer ownership of the form even when the logged in user is not the given form's owner should be mitigated in the controller layer, where checks for whether the current user is the form admin should (and currently is) undertaken.