build: remove babel-loader from webpack dev
When building the frontend using webpack, GA shows a series of
`UnhandledPromiseRejectionWarning` errors related to socket hangups,
and the frontend ultimately fails to build. This happens frequently
in the dev environment, but has never occurred for the production
webpack build.

The only difference between the dev webpack config and the "common"
config shared with prod is that babel-loader is used in dev. So I
attempted to remove babel-loader from dev, since it is not necessary
for the dev environment anyway. This seemed to stop the errors.