build: move end-to-end tests to GitHub Actions (#2353)
* build: move end-to-end tests to GitHub Actions

* build: remove babel-loader from webpack dev

When building the frontend using webpack, GA shows a series of
`UnhandledPromiseRejectionWarning` errors related to socket hangups,
and the frontend ultimately fails to build. This happens frequently
in the dev environment, but has never occurred for the production
webpack build.

The only difference between the dev webpack config and the "common"
config shared with prod is that babel-loader is used in dev. So I
attempted to remove babel-loader from dev, since it is not necessary
for the dev environment anyway. This seemed to stop the errors.

* test(e2e): increase waiting time for download file

Tests were still flaky due to downloaded files (e.g. storage mode CSV)
taking too long to save to disk, resulting in Testcafe timing out and
failing the test. The solution is to increase the waiting time between
downloading the file and checking for its contents.

* build: reduce e2e test concurrency

Tests were still flaky, likely due to Headless Chrome running too slowly.
I finally gave in and reduced the concurrency from 3 to 2 to reduce memory
consumption and hopefully make the tests more reliable.