feat: set INLINE_RUNTIME_CHUNK to false when building React app
The backend uses CSP to prevent inline scripts in HTMLs.

By default, Create React App will embed the runtime script into index.html during the production build. When set to false, the script will not be embedded and will be imported as usual. This will allow express to load the script as usual