fix(ci/deploy): wrap string to truncate in quotes, lower char limit
We use git commit messages as the description for our ELB deployment.
AWS has a hard limit of 200 characters, which may exceed when a commit message is longer (like this) and throw an error. This commit attempts to fix that by truncating the commit message if it is longer than 180 characters.