fix: allow for custom form shape; rerender drawer using key
This commit contains a couple of significant changes.

First, useEditFieldForm is updated to take a transform prop, which
allows the shape of the form inputs to be different from the field.
This allows e.g. for checkbox field options to be read as a string
from the form, then transformed into a string[] in the field.

Second, the useEffect in useEditFieldForm is removed. Its purpose
was to ensure that the edit field drawer correctly rerendered when
the field being edited changed, but it was implicitly relying on
useDebounce to be called first in order to work properly. Instead,
a (still hacky but) more explicit implementation using component
keys was chosen, which is explained in the code comments.