fix: ensure clicking field opens to field editor
Previously, when the builder drawer was closed, clicking on a
form field would open the drawer to the field list instead of
directly to the editing page for that field. This was because
of a useEffect which set fields to inactive whenever the active
tab changed.

Instead, we want to set the fields to inactive only when the active
tab is not the builder.