feat(v2): edit checkbox field in form builder (#3581)
* refactor: generalise EditFieldProps type

* refactor: use SectionFieldBase for input props

* refactor: add index.ts to edit-fieldtype

* refactor: extract getButtonText

* refactor: extract edit form logic into custom hook

* chore: fix section title label

* style: extend divider to ends of drawer

Divider between drawer elements was not extending all the way to the
drawer edge as per the design.

* fix: allow for custom form shape; rerender drawer using key

This commit contains a couple of significant changes.

First, useEditFieldForm is updated to take a transform prop, which
allows the shape of the form inputs to be different from the field.
This allows e.g. for checkbox field options to be read as a string
from the form, then transformed into a string[] in the field.

Second, the useEffect in useEditFieldForm is removed. Its purpose
was to ensure that the edit field drawer correctly rerendered when
the field being edited changed, but it was implicitly relying on
useDebounce to be called first in order to work properly. Instead,
a (still hacky but) more explicit implementation using component
keys was chosen, which is explained in the code comments.

* ref: remove unused variable in BuilderAndDesignContent

* ref: pass original field into transform output

* style: remove minW="max-content" for drawer

This CSS property was causing the content of the drawer to wrongly
overflow in the x direction.

* feat: create and render checkbox edit form

* fix: keepValues when resetting form

The form reset was causing a bug where if customMax or customMin
was undefined, the field would be saved correctly, but the UI would
show the customMin or customMax value reverting to its previous value.

* docs: add comment to explain keepValues

* refactor: remove FormShape casts

* refactor: set custom(Min|Max) to empty string on reset