fix(EditLogicBlock): unregister inputs when logic inputs change
this fixes a bug where both `preventSubmitMessage` and the fields to `show` will both be POSTed if the logic types were switched whilst editing the logic.

By unregistering inputs when they are removed from the DOM, the submitted body will be in the expected shape for their respective logic types instead of containing both.