feat: add patch-package to remove `required` prop in useFormControl
Known issue of various screen readers to read out input as "invalid" even when there has not been any action taken by the user. This is tEcHniCaLLy cOrReCt since an empty required field is technically invalid, though it is very unintuitive to the user.

The most correct way is to remove the `required` prop from the input and use `aria-required` instead, but  ChakraUI injects the `required` prop without allowances for overrides.

Thus, require monkey-patching the underlying package to remove the assignment of `required` prop.