refactor(i18next): make translations optional
Allow translations to be optional, so that development is not
held back by waiting for translations.

Previously, these were required, so either the team waits for
translated text, or is forced to use placeholders. This either slows
the team down, or litters the codebase with unhelpful locale text
to be laboriously translated later.

- Allow features within Translation locales to be optional
  - Extend Translation to FallbackTranslation, which still makes them
    mandatory
- Declare en-sg as a FallbackTranslation
- Trim zh-sg to only have `login` translations, removing all other
  placeholder text