fix(config): default S3 endpoint with region
`AWS_ENDPOINT` defaults to an S3 endpoint hard-coded to ap-southeast-1.
This affects deployment scenarios to other AWS regions, unless
`AWS_ENDPOINT` is specified to line up with `AWS_REGION`.

It can be argued that the default for `AWS_ENDPOINT` should be driven
by the specific region that FormSG is deployed to, so accordingly, at
runtime, change `loadBucketS3UrlSchema()` so that the default endpoint
for S3 buckets reflects the region where FormSG is currently deployed.